PE #14 in progress
Using readlines(), the list was coming in with the \n as a specific character in addition to separating the list elements with commas, which is not helpful. I am moving back toward my original approach. I removed some of my previous non-functioning work because the screen was getting too crowded.